requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
sqlalchemy
pytest
SQLAlchemy~=2.0.20
pyodbc~=4.0.39
python-dotenv-vault~=0.6.3